refactor, add a parser layer and ir

+++ ms-runtime/src/version.rs
@@ -1,8 +1,8 @@
 // version.rs
 
-pub const VERSION_MAJOR: u32 = 0;
-pub const VERSION_MINOR: u32 = 1;
-pub const VERSION_PATCH: u32 = 0;
+pub const VERSION_MAJOR: u8 = 0;
+pub const VERSION_MINOR: u8 = 1;
+pub const VERSION_PATCH: u8 = 0;
 
 pub fn version() -> String {
     format!("{}.{}.{}", VERSION_MAJOR, VERSION_MINOR, VERSION_PATCH)
